Final implementation of zr and several bugfixes
Changed the implementation for zr. Since zr can be found through the negation of Or over all bits, I have used 2 8-way Or gates as well as one 2-way Or gate. The final result is then negated to get zr.

Added several subbuses to the no function, to use them as inputs for zr and ng.